structure the code a bit
