+++ ddpui/routes.py
@@ -24,7 +24,6 @@ from ddpui.api.dashboard_native_api import dashboard_native_router
 from ddpui.api.filter_api import filter_router
 from ddpui.api.public_api import public_router
 from ddpui.api.report_api import report_router
-from ddpui.api.comments_api import comments_router
 
 
 src_api = NinjaAPI(
@@ -88,7 +87,6 @@ dashboard_native_router.tags = ["Native Dashboards"]
 filter_router.tags = ["Filters"]
 public_router.tags = ["Public"]
 report_router.tags = ["Reports"]
-comments_router.tags = ["Comments"]
 
 # mount all the module routes
 src_api.add_router("/api/airbyte/", airbyte_router)
@@ -110,7 +108,6 @@ src_api.add_router("/api/charts/", charts_router)
 src_api.add_router("/api/dashboards/", dashboard_native_router)
 src_api.add_router("/api/filters/", filter_router)
 src_api.add_router("/api/reports/", report_router)
-src_api.add_router("/api/comments/", comments_router)
 
 # Public API without authentication
 public_api = NinjaAPI(
